Add check for ZK node empty data
Because we have a pattern of caching a list of node IDs and then
iterating through them, retrieving the data for each, it's possible
our cache could get stale and the node we try to retrieve is no longer
valid. In that case, we could get empty data for the node. This adds
a check for that.

Change-Id: I7c6c396520ae324fef9c6b92e8f34a42128f5737