Add test config with min-ready of 0
Tests that validate statsd reporting, and that create more than one
node, have a race on what is reported to statsd. In rare cases, both
nodes can become READY before the first report to statsd, so we'd not
get an expected '1' value for 'nodepool.nodes.ready', but get a '2'
instead. This adds a config that doesn't create the min-ready node so
we can eliminate the race.

Change-Id: I4717a2ff4313dd8d21a74e085d730c5d254c12c6