Fix test_node_vhd_and_qcow2 to validate uploads
The test was assuming that the 2 built nodes would be evenly
satisfied across providers. We cannot guarantee that for v3
nodepool. We CAN guarantee that the images that are built are
uploaded to the provider(s) that supports the image types. In
this case, fake-provider1 supports vhd and fake-provider2
supports qcow2. This required adding a 'format' field to the
ImageUpload model so we can know which format was uploaded.

Change-Id: I7db211b511abbb82807b50c786fa19a25ae10d32