Add image-id and image-name options to cloud-images
The cloud-image name is currently used both to specify the image
in the cloud, and also as a cross-referencing key within the
nodepool config.  As such, it ends up being repeated within the config
(possibly quite often in large configurations).

Separate these functions so that an image can be identified once in
a cloud provider, and referenced from multiple labels with the internal
key.  This makes for improved readability in some cases (such as long
cloud image names, or specifying images by uuid), and reduces churn
when cloud image identifiers change.

Change-Id: I83f2902be4b9b73a949461b7f14da548066b9562