Don't join image upload workers on stop()
Image uploads are synchronous with their workers and because writing
logs of bytes to clouds can be slow this means that join()ing on an
image upload worker can take a significant amount of time.

When a nodepool builder receives a sigint we want it to stop in a
reasonable amount of time and gracefully close connections to the
zookeeper database so that locks are released properly and records can
be cleaned safely. The old stop() code which handles sigint joined on
the upload worker threads. This meant it couldn't happen in a reasonable
amount of time for the reason above. This then leads to killing the
process in init scripts with sigkill.

Thankfully we can just not wait for upload workers to join and let
process exit kill the upload process for us. Separately we can
gracefully close the zookeeper connection. Then any other builders
(possibly when this one restarts) can clean up the upload record in zk
and in the cloud.

Change-Id: I52425bb8e5b8f0d6e1d25674cbe590e32b629e6d