Remove the default argument to secure in nodepoolcmd
The nodepool secure file is currently not being used for anything. By
having a default value in the nodepool command the config will try to
load this file. When the file then doesn't exist the program will exit.

This basically means you have to have an empty file at
/etc/nodepool/secure.conf to be able to use nodepoolcmd. Launcher does
not specify a default for secure, so follow that lead and remove the
default.

Change-Id: I244f097bcd1443588c03257394c5e8768ab3e785
Signed-off-by: Jamie Lennox <jamielennox@gmail.com>