Improve exception handling around lost requests
An unexpected exception during resets of lost requests was leaving
the request in a locked state, and thus it would never get processed
again. Improve handling around that.

Also, raise a ZKLockException if the thing we've been asked to lock
has disappeared for some reason (this was the unhandled exception that
caused the lost request handling to fail).

Change-Id: Ie3e91714edc482b7b4fb99d7992cae999b1b7026