Harden exception handling around instance deletes
It's possible to leave a ZK node in a locked state if deleting an
instance fails for some reason. This hardens the logic in the cleanup
thread to make sure we always attempt to unlock the node if the delete
fails.

We have a similar situation in clearing the allocation of nodes whose
request has disappeared.

Also, s/unallocate/deallocate/

Change-Id: I82d9dcb0f86c3296f2427a315e14fce772f5369a