Unpause a declined request
A paused request is retried until it can be satisfied. If conditions
change such that, during a retry, we now determine we should decline
the request, then we need to unpause it so it can be removed from our
active request handlers. Without doing this, we would have unallocated
any previously allocated nodes for the request, leaving things in a
state that could cause chaos and agony.

Change-Id: I9073ca2274ef4e97ecb1ece01ddb861c22807984