Be consistent with the ZK data model
An earlier change added a non-None default value to the Node class
in order to fix tests, but this is inconsistent with the rest of
the model where we use None for the defaults. This can be fixed
by always storing the username, regardless of value, which is
actually more consistent, too.

Change-Id: If420ddb8de4c0b7ae5ff2edf52d3b0230f2da126