Reorg non detailed instance listing columns
"AZ" is an almost always unused piece of info so move it from non
detailed info to detailed info when listing instances. Public IP addrs
are used all the time though so move from from detailed to non detailed
listings.

This should give better balance to the default usage of the list command
for day to day ops.

Change-Id: I5bb6255ea76b78b8427d1f145d3f8e077f11013d