Handle invalid request lock in request cleanup
We have seen in production the case where node request handler cleanup
fails due to being unable to cleanup a node request lock. This appears
to happen because the node request lock has been invalidated somehow.
The lock may or may not be valid in zookeeper, but the python object has
been updated to None.

This is a defenseive change to handle when that node request lock is
None. This occurs when the node request has already been removed from
zookeeper so it is safe to ignore this node request lock and move on
with processing other requests in the handler. The node request lock
clenaup step should clean out any leaks on the zookeeper side of things.

Change-Id: I30e78bc67906d9ad97bf9ca964ea145c67526a8b