Log unknown providers during quota calculation
When determining quota usage for known nodes log cases where the
provider config isn't known. One theory is that this can happen when
launchers are restarted and that it is otherwise normal but without more
information it is hard to know for sure. Log the node data when we hit
this sutation so that we can dig in more and debug this.

Change-Id: I466e64a999ebf867ada9462a0c00a90de4b70125