Do not attempt to handle requests when disabled
The PoolWorker thread will just spam the logs by continually
declining requests when it is disabled with max-servers. This
prevents the thread from trying to do that anymore.

This also adjusts the max_servers quota short-circuit in the
OpenStack handler to allow values less than 0 to disable it.
OpenStack Infra has used -1 for this in the past, so seems like
a good idea to continue to allow its use.

Change-Id: Ia2956dea528f887269ca0f4507d6e7cb902c3b6c