Fix race in test_failed_provider
In test_failed_provider we determine what the second node that
fake-provider2 booted is then delete it. Unfortunately there is a race
where we could find the first node booted in this provider and re delete
it instead of also deleting the second node which never frees up
provider quota which cases the test to timeout and fail.

We fix this by explicitly checking the node state when looking for the
second node booted in the provider. The first node will be in a deleting
state or just won't exist so checking for a ready node should be
sufficient to always find the second node.

We also fix a small bug in how we raise the exception that causes
fake-provider's launch request to fail. We need a single argument to the
monkey patched launch() function otherwise we get a TypeError from
python instead of the KeyError that we want.

Change-Id: Ia2fad9736be76d6ca9a923c8e0643a96a9d1dbe6