Partial revert for disabled provider change
Commit 1cf2cddf9ded2641a53977e6a9e6f0a24cd062c9 attempted to
short-circuit request handling when a provider was disabled by
setting max-servers to 0. That prevents the provider from declining
any requests, and thus prevents a request from ever reaching
FAILURE status if none of the other providers can satisfy it (the
request declined_by array is compared against the list of registered
launchers).

This adds a test for this case (which was run against the current
code and proved it was broken).

Change-Id: I8a2f9771b55b08259ecffa3475bdd2d4f2654464