Fix nodepool-zuul-functional
Several fixes to this job:

We need to run bindep twice - once for nodepool, once for zuul. Add
invocation of bindep role and also copy install-distro-packages so that
the job works, next step is remove the install-distro-packages from
zuul.

Add a post-run to copy the nodepool logs, so you can diagnose what's
going wrong if the jobs fails

Fix up a configuration issue, it tries to write build-logs to
/var/log/nodepool which it doesn't own, redirect to the temp area.

Add it as a non-voting check job

Depends-On: https://review.openstack.org/545163
Change-Id: I12db55d3e4c7a71b9af56567858df0a620ee3b73