Only DeletedNodeWorker should delete nodes
The proper/expected way to delete a node is to set its state to
DELETING and then allow the DeletedNodeWorker to actually delete
the instance and zookeeper node. Not doing it this way introduces
the race seen in commit 59d636740a00f0a8157e73a606b65fd8076193ca.
This change fixes all the places (other than the nodepool command
to delete --now) where we were not doing it properly.

Change-Id: I1962ad672d3d5e646ed29beb07b720fc53d604ed