Add additional builder debug logging
Tracking down builder decisions to do (or not do) things is made
difficult by not having insight into certain data. Particularly,
the upload recency table is used to determine whether or not images
should be deleted from a provider. Having that data in debug logs
helps tremendously.

We also log when we delete the upload znode record and remove the
'stats' portion of ImageUpload str representation because it turns
out to not be as useful as I had hoped in the beginning.

Change-Id: Ic6e1ea0811879ed8eaf0216ef2a3b68e9380e727