Revert "Add /label-list to the webapp"
This reverts commit b7378892502836f257456cd277190ee75a7ca75f.

As discussed in [1] the label list is intended to provide a list of
available labels so users don't have to dig into configuration files
to determine what is available to run jobs on.  As written it possibly
misses nodes without a min-ready or in-use nodes, so an alternative
implementation is probably called for.  This is also currently returns
values in a plain list, rather than a list of dicts, which prevents
some de-deuplication refactoring.

Thus remove it for now, pending some more work.

[1] http://lists.zuul-ci.org/pipermail/zuul-discuss/2018-February/000039.html

Change-Id: Ic88da24dd5d140697d4af8e6f5ee0936c44f59c5