Handle ZK session loss during node launch
Node launches can take a while, thus increasing the chances of it
encountering a lost connection to ZooKeeper, which could lead to
a session loss. Let's try to handle that by not re-attempting any
launches (pointless because we've lost our node lock and the node
could be deleted out from under us).

Change-Id: Iaca005355e8d03e1e5ab7cdab0ea3c04117ca6fb