Reduce strictness of growroot test
We are seeing consistent failures where the Trusty root partition is
64MiB shorter than we expect.  Unfortunately we don't currently have a
concrete explanation for this but rounding due to alignment is
suspected.

Reduce the expected size to something bigger than the images, but not
so close to 5GiB.  Also add a more useful failure message; currently
you have to dig back through the logs to see where it went wrong.

Change-Id: Iba1fafdb1fe0f3c1b751772af939f079c429fcf3