Fix race in test_hold_expiration_no_default
The test_hold_expiration_no_default test was misconfigured to use
the nodepool config for the max ready age tests, which was causing
a race in the test where the node could be removed (b/c it exceeded
the max ready age) before it could set a hold. This fixes it to use
a new config, and lowers the TTL on the held node to quicken the test.

Change-Id: Iaeb16e7a11f8b5c7bb2b363b7939fee7933393c4