Add systemd drop-in file for CentOS 7
The existing service file is for Fedora27 and it does not work
on CentOS 7 since pip installs nodepool in /usr/bin on CentOS 7.
This change adds the systemd drop-in file for this distro so
ExecStart command is overriden by systemd based on where nodepool
is installed.

This change is needed in order to have complete installation instructions
for Centos7 in zuul-from-scratch document.

Change-Id: Id6a6303d4785b64d9d259b9d553dea6d07eeeedb
Signed-off-by: Fatih Degirmenci <fdegir@gmail.com>