Fix test patching of clouds.yaml file locations
OpenStack Client Config has been pulled into openstacksdk. As part of
this work OSCC internals were dropped and aliased into the sdk lib. This
move broke patching of the clouds.yaml file location for nodepool tests.

We quickly work around this by using the new location for the value to
be overridden in openstacksdk.

Change-Id: I55ad4333ffddec8eeb023e345156e96773504400