Add openstacksdk to nodepool-functional-py35-src
openstacksdk is in the critical path now. Add it to the src job so that
we can add the nodepool-functional-py35-src job to openstacksdk and make
sure we're not breaking nodepool when we change things. We also need
this job on the shade repo, but that should be good to go already.

Change-Id: If11680dba61fe81997e8f3083c9db60d37a34ccc