Default image creation to qcow2 type
It is described in the docs, that nodepool will build image only if it
is referred in one of the clouds, but it is possible to force that
creation: `nodepool image-build fedora-28`.

This leads to inability to define image types and an empty "-t"
argument to DIB; which causes the command invocation to fail.

Default the type value to qcow2.  This matches the default type that
dib will create and keeps with the general principle of nodepool
figuring out what type of image to make for you.

Change-Id: I27f8f45ba62e5ee0cca0ea7cd27893092b726ebd
Story: 2001963