Make manage_images a property of ProviderConfig
The current driver API does not make it clear that a driver must
set the 'manage_images' property on the driver of the ProviderConfig
to indicate that it can manage external images. Since this is really
a function of the provider, not the driver, make this an abstract
property on the ProviderConfig that must be defined by each driver
implementation.

Change-Id: Iefe257f943aaa5740d0326ca6150632c21a2a8cf