Add multilabel support to ZooKeeper Node model
Support storing multiple node types/labels for a Node() in the
ZooKeeper data.

Neither the OpenStack nor the static drivers make use of this yet;
they assume there will only be one value in the type/label list.

In the request handler code, we must now store the satisfied types
(rather than rely on the Node.type in self.nodeset) since that can
contain multiple values and we can't tell which value was used to
satisfy the request. The LabelRecorder class helps do this, and to
also help with assigning nodes to a request in proper order based
on order of the labels in the request.

Change-Id: Ice46e22183c4e7354eaa66504ea147aff5ee0df3