Fix adding qcow2 format without need
Currently nodepool is defaulting to the qcow2 if no format was given
in the diskimage config. However nodepool also can auto-detect the
image format based on clouds.yaml. This is not taken into account
leading to qcow2 added when using raw-only clouds and auto-detection.

This can be fixed by delaying the fallback to the point where we
actually do the build.

Change-Id: Ic1d9b5abd837a3225cc13da30a80ebf1dda61b64