Pre-register static nodes
This changes the static driver to register nodes from its config
file with ZooKeeper at startup and when the config file changes.
A node is registered max-parallel-jobs times with ZooKeeper (default
is 1).

Once a node is deleted, it is immediately re-registered, unless it
no longer exists in our config, or it's not needed based on a change
to max-parallel-job count.

Change-Id: Ie386e440c2016af390087d109f7e5211c4fd968d