Clarify static node name uniqueness in docs
Static node names are either a hostname or IP address. By their
very nature, this needs to be unique across all nodes defined in
the config, even across pools. And the static driver assumes this
is the case. Add text to make sure this is explicitly pointed out.

Also, note that multiple pools are supported, but useless.

Change-Id: I8c8c1e2cda9b0250910c4af455b8e1472ad86151