Fix race at shutdown
Currently we first stop the zk connection and afterwards the
providers. This leads to a race condition where the providers may
still want to communicate with zk while the connection is already
dead. Instead stop the zk connection after stopping the providers.

This also fixes some occational and misleading zk connection
exception logs at the end of test cases.

Change-Id: Ibb13f74a1130190403dd79922ccd6c2b75daffc9