Use zk connection passed to OpenStack driver
We are passing the zk connection to the estimatedNodepoolQuotaUsed()
method of the OpenStack provider when we already have it passed into
the provider's start() method, but not saving it there. Let's save
that connection in start() and use it instead.

In a later change to the provider, we will make further use of it.

Change-Id: I013a28d6c46046497d8b04867c51a23f6fa49d39