Ensure that completed handlers are removed frequently
On a busy system it can happen that assignHandlers takes quite some
time (we saw occurrences of more than 10 minutes). Within this time no
node request is marked as fulfilled even if the nodes are there. A
possible solution is to return from assignHandlers frequently during
the iteration so we can remove completed handlers and then proceed
with assigning handlers.

Change-Id: I10f40504c81d532e6953d7af63c5c58fd5283573