Alter the metasyntactic variable in driver docs
We're trying to convey a hierarchy of options which may change
depending on what options are selected.  Given the config:

providers:
  - driver: openstack
    labels: [foo]

Rather than naming the option "providers.openstack.labels", name
it "providers.[openstack].labels" to indicate that the "openstack"
part of the option is not a normal name, but metasyntactic.  Use
square brackets to suggest it's an item in a list.

Also some minor style updates, and add 'kubernetes' to the list
of driver values.

Change-Id: I33db936475eaac207b722a9dbfcd8abe99cfc7ed