Update node request during locking
After locking a node request we most of the time update ot to double
check if the state changed on us. This is important to not operate on
stale data as we only have a guarantee that the node request data is
updated if we update the data after getting the lock (regardless if we
operate on cached data or not). However it is cumbersome to do this
after every lockNodeRequest call so doing this centrally reduces the
risk to forget that.

This does not introduce any behavior change.

Change-Id: Ie6713fdd01225bbcff7839605ac9f94dc6cb2f78