OpenStack: count leaked nodes in unmanaged quota
If a node has leaked, it won't be counted against quota because
it's still recognized as belonging to the nodepool provider so it
doesn't count against unmanaged quota, however, there is no zk
record for it, so it also isn't counted against managed quota.
This throws quota calculations off for as long as the leaked
instances exist in nova.

To correct this, count leaked nodes against unmanaged quota.

Change-Id: I5a658649b881ed80b777096ec48cb6207f2a9cc6