OpenStack: store ZK records for launch error nodes
If we get an error on create server, we currently leak the instance
because we don't store the external id of the instance in ZK.  It
should eventually be deleted since it's a leaked instance, but
we try to keep track of as much as possible.  OpenStackSDK can
often return the external id to us in these cases, so handle that
case and store the external id on a ZK record so that the instance
is correctly accounted for.

Change-Id: I7ec448e9a7cf6cd01903bf7b5bf4b07a1c143fb8